refactor: Refactor RequestOverrider
This makes RequestOverrider into a class and breaks its logic up into several shorter methods. There is more that can be done here, though this seems like a meaningful place to pause.
