+++ lib/intercept.js
@@ -4,7 +4,7 @@
  * @module nock/intercepts
  */
 
-const RequestOverrider = require('./request_overrider')
+const { InterceptedRequestHandler } = require('./intercepted_request_handler')
 const common = require('./common')
 const { inherits } = require('util')
 const Interceptor = require('./interceptor')
@@ -275,7 +275,12 @@ function overrideClientRequest() {
       debug('using', interceptors.length, 'interceptors')
 
       //  Use filtered interceptors to intercept requests.
-      const overrider = RequestOverrider(this, options, interceptors, remove)
+      const overrider = new InterceptedRequestHandler(
+        this,
+        options,
+        interceptors,
+        remove
+      )
       Object.assign(this, overrider)
 
       if (callback) {
